refactor: extract common code from SearchPage.test.tsx
This then allows two things:

1. It to be used in a new .test.tsx file for testing advanced search;
and
2. Allow us to start splitting up the monolithic SearchPage.test.tsx to
address slow execution times.
